main.py
import pandas as pd
import numpy as np
from sklearn.model_selection import train_test_split
from sklearn.ensemble import RandomForestRegressor
from datetime import datetime, timedelta
import requests
import os
from dotenv import load_dotenv

class SportsAnalyzer:
    def __init__(self):
        self.data = None
        self.model = RandomForestRegressor(n_estimators=100)
        
    def fetch_player_data(self, player_name, sport="NBA"):
        """
        Fetch historical player data
        Note: You'll need to implement this with a proper sports data API
        """
        # Placeholder for API implementation
        # You can use APIs like SportRadar, ESPN, or NBA Stats
        pass

    def preprocess_data(self, data):
        """
        Preprocess player statistics for analysis
        """
        # Convert data to pandas DataFrame
        df = pd.DataFrame(data)
        
        # Add features like:
        # - Rolling averages (last 5, 10 games)
        # - Home/Away game indicator
        # - Days of rest
        # - Opponent strength
        
        return df

    def train_model(self, feature_columns):
        """
        Train the prediction model using historical data
        """
        X = self.data[feature_columns]
        y = self.data['actual_points']  # or whatever metric we're predicting
        
        X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.2)
        self.model.fit(X_train, y_train)
        
        return self.model.score(X_test, y_test)

    def predict_performance(self, player_name, metric):
        """
        Predict player performance for a specific metric
        """
        # Get recent player data
        recent_data = self.fetch_player_data(player_name)
        processed_data = self.preprocess_data(recent_data)
        
        prediction = self.model.predict(processed_data)
        
        return {
            'player': player_name,
            'metric': metric,
            'predicted_value': prediction[0],
            'confidence_score': self.model.score(processed_data, recent_data['actual_points'])
        }

    def get_recommendation(self, player_name, metric, line):
        """
        Get over/under recommendation based on predicted performance
        """
        prediction = self.predict_performance(player_name, metric)
        
        confidence_threshold = 0.7
        if prediction['confidence_score'] < confidence_threshold:
            return "Not enough confidence to make a prediction"
        
        if prediction['predicted_value'] > line * 1.1:  # 10% margin
            return "OVER"
        elif prediction['predicted_value'] < line * 0.9:  # 10% margin
            return "UNDER"
        else:
            return "TOO CLOSE - AVOID"

def main():
    analyzer = SportsAnalyzer()
    
    # Example usage
    player_name = "LeBron James"
    metric = "points"
    line = 25.5
    
    recommendation = analyzer.get_recommendation(player_name, metric, line)
    print(f"Recommendation for {player_name} {metric} (line: {line}): {recommendation}")

if __name__ == "__main__":
    main()
