requirements.txt
pandas==2.0.3
requests==2.31.0
python-dotenv==1.0.0
scikit-learn==1.3.0
numpy==1.24.3
beautifulsoup4==4.12.2
click==8.1.3
tabulate==0.9.0
streamlit==1.24.0
plotly==5.15.0
lxml==4.9.3
